reorganize structure for new year 2021
