+++ README.md
@@ -6,48 +6,63 @@ Advent of code - personal solutions based of the following website.
 
 Please have a look yourself, and I encourage you to solve listed puzzles on your own.
 
-## Day 1
+## Table of content
+
+- [Year 2020](#year-2020)
+- [Year 2021](#year-2021)
+
+## Year 2020
+
+### Day 1: Report Repair
 
 > todo..
 
-## Day 2: Password Philosophy
+### Day 2: Password Philosophy
 
 Solved puzzle.
 
 - https://adventofcode.com/2020/day/2
 
-## Day 3: Toboggan Trajectory
+### Day 3: Toboggan Trajectory
 
 Solved puzzle.
 
 - https://adventofcode.com/2020/day/3
 
-## Day 4: Passport Processing
+### Day 4: Passport Processing
 
 Solved puzzle.
 
 - https://adventofcode.com/2020/day/4
 
-## Day 5: Binary Boarding
+### Day 5: Binary Boarding
 
 Solved puzzle.
 
 - https://adventofcode.com/2020/day/5
 
-## Day 6: Custom Customs
+### Day 6: Custom Customs
 
 Solved puzzle.
 
 - https://adventofcode.com/2020/day/6
 
-## Day 7: Handy Haversacks
+### Day 7: Handy Haversacks
 
 Solved puzzle.
 
 - https://adventofcode.com/2020/day/7
 
-## Day 8: Handheld Halting
+### Day 8: Handheld Halting
 
 Solved puzzle.
 
 - https://adventofcode.com/2020/day/8
+
+## Year 2021
+
+### Day 1: Sonar Sweep
+
+Solved puzzle.
+
+- https://adventofcode.com/2021/day/1
