--- src/main/java/com/github/anjeyy/adventcode/eight/Operation.java
@@ -1,19 +1,0 @@
-package com.github.anjeyy.adventcode.eight;
-
-import java.util.Arrays;
-
-enum Operation {
-    ACCUMULATOR("acc"),
-    JUMPS("jmp"),
-    NO_OPERATION("nop");
-
-    private final String abbreviation;
-
-    Operation(String abbreviation) {
-        this.abbreviation = abbreviation;
-    }
-
-    static Operation from(String string) {
-        return Arrays.stream(values()).filter(s -> s.abbreviation.equals(string)).findAny().orElseThrow();
-    }
-}
