--- src/main/java/com/github/anjeyy/adventcode/two/Password.java
@@ -1,31 +1,0 @@
-package com.github.anjeyy.adventcode.two;
-
-import com.github.anjeyy.adventcode.Constants;
-
-class Password {
-
-    private final String phrase;
-
-    private Password(String phrase) {
-        this.phrase = extractPasswordPhrase(phrase);
-    }
-
-    String getPhrase() {
-        return phrase;
-    }
-
-    static Password from(String phrase) {
-        return new Password(phrase);
-    }
-
-    private String extractPasswordPhrase(String phrase) {
-        return phrase.split(Constants.WHITESPACE)[2];
-    }
-
-    @Override
-    public String toString() {
-        return "Password{" +
-            "phrase='" + phrase + '\'' +
-            '}';
-    }
-}
