+++ src/main/java/com/github/anjeyy/adventcode/year2020/four/EyeColor.java
@@ -1,0 +1,21 @@
+package com.github.anjeyy.adventcode.year2020.four;
+
+enum EyeColor {
+    AMB("amb"),
+    BLUE("blu"),
+    BROWN("brn"),
+    GRAY("gry"),
+    GREEN("grn"),
+    HAZEL("hzl"),
+    OTHER("oth");
+
+    private final String colorAbbreviation;
+
+    EyeColor(String colorAbbreviation) {
+        this.colorAbbreviation = colorAbbreviation;
+    }
+
+    public String getColorAbbreviation() {
+        return colorAbbreviation;
+    }
+}
